Improve pages router deployment skew tests
